+++ Sources/PopupBridge/NetworkError.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+enum NetworkError: Error {
+    case invalidResponse
+}