+++ Sources/PopupBridge/POPPopupBridge.swift
@@ -8,13 +8,17 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler {
     /// Exposed for testing
     var returnedWithURL: Bool = false
     
+    static var analyticsService: AnalyticsServiceable = AnalyticsService()
+    
     // MARK: - Private Properties
     
     private let messageHandlerName = "POPPopupBridge"
-    private let hostName = "popupbridgev1"    
+    private let hostName = "popupbridgev1"
+    private let sessionID = UUID().uuidString.replacingOccurrences(of: "-", with: "")
     private let webView: WKWebView
-    private var webAuthenticationSession: WebAuthenticationSession = WebAuthenticationSession()
+    private let application: URLOpener = UIApplication.shared
     
+    private var webAuthenticationSession: WebAuthenticationSession = WebAuthenticationSession()
     private var returnBlock: ((URL) -> Void)? = nil
     
     // MARK: - Initializers
@@ -29,24 +33,33 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler {
         self.webView = webView
         
         super.init()
-
+        
+        Self.analyticsService.sendAnalyticsEvent(PopupBridgeAnalytics.started, sessionID: sessionID)
+        
         configureWebView()
         webAuthenticationSession.prefersEphemeralWebBrowserSession = prefersEphemeralWebBrowserSession
                 
-        returnBlock = { url in
-            guard let script = self.constructJavaScriptCompletionResult(returnURL: url) else {
+        returnBlock = { [weak self] url in
+            guard let script = self?.constructJavaScriptCompletionResult(returnURL: url) else {
                 return
             }
 
-            self.injectWebView(webView: webView, withJavaScript: script)
+            self?.injectWebView(webView: webView, withJavaScript: script)
             return
         }
     }
 
     /// Exposed for testing
-    convenience init(webView: WKWebView, webAuthenticationSession: WebAuthenticationSession) {
+    convenience init(
+        webView: WKWebView,
+        webAuthenticationSession: WebAuthenticationSession
+    ) {
         self.init(webView: webView)
         self.webAuthenticationSession = webAuthenticationSession
+    }
+    
+    deinit {
+        webView.configuration.userContentController.removeAllScriptMessageHandlers()
     }
     
     // MARK: - Internal Methods
@@ -76,10 +89,12 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler {
 
         if let payloadData = try? JSONEncoder().encode(payload),
            let payload = String(data: payloadData, encoding: .utf8) {
+            Self.analyticsService.sendAnalyticsEvent(PopupBridgeAnalytics.succeeded, sessionID: sessionID)
             return "window.popupBridge.onComplete(null, \(payload));"
         } else {
             let errorMessage = "Failed to parse query items from return URL."
             let errorResponse = "new Error(\"\(errorMessage)\")"
+            Self.analyticsService.sendAnalyticsEvent(PopupBridgeAnalytics.failed, sessionID: sessionID)
             return "window.popupBridge.onComplete(\(errorResponse), null);"
         }
     }
@@ -87,12 +102,16 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler {
     /// Injects custom JavaScript into the merchant's webpage.
     /// - Parameter scheme: the url scheme provided by the merchant
     private func configureWebView() {
-        webView.configuration.userContentController.add(self, name: messageHandlerName)
+        webView.configuration.userContentController.add(
+            WebViewScriptHandler(proxy: self),
+            name: messageHandlerName
+        )
         
         let javascript = PopupBridgeUserScript(
             scheme: PopupBridgeConstants.callbackURLScheme,
             scriptMessageHandlerName: messageHandlerName,
-            host: hostName
+            host: hostName,
+            isVenmoInstalled: application.isVenmoAppInstalled()
         ).rawJavascript
         
         let script = WKUserScript(
@@ -139,7 +158,9 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler {
                             window.popupBridge.onComplete(null, null);\
                         }
                     """
-
+                    
+                    Self.analyticsService.sendAnalyticsEvent(PopupBridgeAnalytics.canceled, sessionID: sessionID)
+                    
                     injectWebView(webView: webView, withJavaScript: script)
                     return
                 }
@@ -153,13 +174,8 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler {
 extension POPPopupBridge: ASWebAuthenticationPresentationContextProviding {
 
     public func presentationAnchor(for session: ASWebAuthenticationSession) -> ASPresentationAnchor {
-        if #available(iOS 15, *) {
-            let firstScene = UIApplication.shared.connectedScenes.first as? UIWindowScene
-            let window = firstScene?.windows.first { $0.isKeyWindow }
-            return window ?? ASPresentationAnchor()
-        } else {
-            let window = UIApplication.shared.windows.first { $0.isKeyWindow }
-            return window ?? ASPresentationAnchor()
-        }
+        let firstScene = UIApplication.shared.connectedScenes.first as? UIWindowScene
+        let window = firstScene?.windows.first { $0.isKeyWindow }
+        return window ?? ASPresentationAnchor()
     }
 }