+++ UnitTests/PopupBridge_UnitTests.swift
@@ -7,6 +7,7 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
     let scriptMessageHandlerName: String = "POPPopupBridge"
     let returnURL: String = "com.braintreepayments.popupbridgeexample"
     let mockWebAuthenticationSession = MockWebAuthenticationSession()
+    let mockAnalyticsService = MockAnalyticsService()
 
     var webViewReadyBlock: (Void)?
 
@@ -33,9 +34,9 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView)
+        let _ = POPPopupBridge(webView: webView)
 
-        XCTAssertEqual(mockUserContentController.scriptMessageHandler as? POPPopupBridge, pub)
+        XCTAssertNotNil(mockUserContentController.scriptMessageHandler as? WebViewScriptHandler)
         XCTAssertEqual(mockUserContentController.name, scriptMessageHandlerName)
     }
 
@@ -52,7 +53,10 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, webAuthenticationSession: mockWebAuthenticationSession)
+        let pub = POPPopupBridge(
+            webView: webView,
+            webAuthenticationSession: mockWebAuthenticationSession
+        )
 
         mockWebAuthenticationSession.cannedResponseURL = URL(string: "http://example.com/?hello=world")
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
@@ -135,7 +139,10 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, webAuthenticationSession: mockWebAuthenticationSession)
+        let pub = POPPopupBridge(
+            webView: webView,
+            webAuthenticationSession: mockWebAuthenticationSession
+        )
         let mockURL = URL(string: "sdk.ios.popup-bridge://popupbridgev1/return?something=foo&other=bar")!
         mockWebAuthenticationSession.cannedResponseURL = mockURL
 
@@ -161,7 +168,10 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, webAuthenticationSession: mockWebAuthenticationSession)
+        let pub = POPPopupBridge(
+            webView: webView,
+            webAuthenticationSession: mockWebAuthenticationSession
+        )
         let mockURL = URL(string: "sdk.ios.popup-bridge://popupbridgev1/return")!
         mockWebAuthenticationSession.cannedResponseURL = mockURL
 
@@ -181,7 +191,10 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         stubMessage.body = stubMessageBody
         stubMessage.name = stubMessageName
 
-        let pub = POPPopupBridge(webView: WKWebView(), webAuthenticationSession: mockWebAuthenticationSession)
+        let pub = POPPopupBridge(
+            webView: WKWebView(),
+            webAuthenticationSession: mockWebAuthenticationSession
+        )
 
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
 
@@ -204,7 +217,10 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         stubMessage.body = stubMessageBody
         stubMessage.name = stubMessageName
 
-        let pub = POPPopupBridge(webView: WKWebView(), webAuthenticationSession: mockWebAuthenticationSession)
+        let pub = POPPopupBridge(
+            webView: WKWebView(),
+            webAuthenticationSession: mockWebAuthenticationSession
+        )
 
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
 
@@ -270,5 +286,91 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         }
         
         return nil
+    }
+    
+    func testInit_sendsAnalytics() {
+        XCTAssertEqual(mockAnalyticsService.eventCount, 0)
+        POPPopupBridge.analyticsService = mockAnalyticsService
+        
+        let _ = POPPopupBridge(
+            webView: WKWebView(),
+            webAuthenticationSession: mockWebAuthenticationSession
+        )
+
+        XCTAssertEqual(mockAnalyticsService.eventCount, 1)
+        XCTAssertEqual(mockAnalyticsService.lastEventName, PopupBridgeAnalytics.started)
+        XCTAssertNotNil(mockAnalyticsService.lastSessionID)
+    }
+    
+    func testConstructJavaScriptCompletionResult_whenReturnURL_sendsAnalytics() {
+        XCTAssertEqual(mockAnalyticsService.eventCount, 0)
+        let configuration = WKWebViewConfiguration()
+        let mockUserContentController = MockUserContentController()
+
+        configuration.userContentController = mockUserContentController
+
+        let webView = WKWebView(frame: CGRect(), configuration: configuration)
+        POPPopupBridge.analyticsService = mockAnalyticsService
+        let pub = POPPopupBridge(
+            webView: webView,
+            webAuthenticationSession: mockWebAuthenticationSession
+        )
+        let mockURL = URL(string: "sdk.ios.popup-bridge://popupbridgev1/return?something=foo&other=bar")!
+        mockWebAuthenticationSession.cannedResponseURL = mockURL
+
+        let expectedResult = "window.popupBridge.onComplete(null, {\"path\":\"\\/return\",\"queryItems\":{\"other\":\"bar\",\"something\":\"foo\"}});"
+        let result = pub.constructJavaScriptCompletionResult(returnURL: mockURL)
+
+        let expectedJSON = extractJSON(from: expectedResult)
+        let actualJSON = extractJSON(from: result!)
+        
+        XCTAssertEqual(actualJSON, expectedJSON)
+        XCTAssertEqual(mockAnalyticsService.eventCount, 2)
+        XCTAssertEqual(mockAnalyticsService.lastEventName, PopupBridgeAnalytics.succeeded)
+        XCTAssertNotNil(mockAnalyticsService.lastSessionID)
+    }
+    
+    func testPopupBridge_whenCancelButtonTappedOnSafariViewController_sendsAnalytics() {
+        XCTAssertEqual(mockAnalyticsService.eventCount, 0)
+        let stubMessageBody: [String: String] = ["url": "http://example.com/?hello=world"]
+        let stubMessageName = scriptMessageHandlerName
+        let stubMessage = MockScriptMessage()
+        stubMessage.body = stubMessageBody
+        stubMessage.name = stubMessageName
+
+        let configuration = WKWebViewConfiguration()
+        let mockUserContentController = MockUserContentController()
+
+        configuration.userContentController = mockUserContentController
+
+        POPPopupBridge.analyticsService = mockAnalyticsService
+        let webView = WKWebView(frame: CGRect(), configuration: configuration)
+        let pub = POPPopupBridge(
+            webView: webView,
+            webAuthenticationSession: mockWebAuthenticationSession
+        )
+        mockWebAuthenticationSession.shouldCancel = true
+        pub.userContentController(WKUserContentController(), didReceive: stubMessage)
+
+        webView.evaluateJavaScript("""
+        "if (typeof window.popupBridge.onCancel === 'function') {"
+        "  window.popupBridge.onCancel();"
+        "} else {"
+        "  window.popupBridge.onComplete(null, null);"
+        "}"
+        """)
+        
+        XCTAssertEqual(mockAnalyticsService.eventCount, 2)
+        XCTAssertEqual(mockAnalyticsService.lastEventName, PopupBridgeAnalytics.canceled)
+        XCTAssertNotNil(mockAnalyticsService.lastSessionID)
+    }
+    
+    func testPOPPopupBridge_shouldNotLeak() {
+        let webView = WKWebView()
+        trackForMemoryLeak(instance: POPPopupBridge(
+            webView: webView,
+            webAuthenticationSession: mockWebAuthenticationSession
+        ))
+        webView.load(URLRequest(url: URL(string: "some-popup-bridge-example")!))
     }
 }