+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Bump minimum supported deployment target to iOS 16+
 * Require Xcode 16.2.0+ and Swift 5.10+
 * Add the `prefersEphemeralWebBrowserSession` property to the `POPPopupBridge` initializer, which specifies whether to request a private authentication session from the browser.
+* Add validation to check if the Venmo app is installed on the device.
 
 ## 2.2.0  (2025-02-05)
 * Require Xcode 15.0+ and Swift 5.9+ (per [App Store requirements](https://developer.apple.com/news/?id=khzvxn8a))